Allow forcing tags to be reported as trending
Using a comma-separated list in the evironment variable
`ALWAYS_TRENDING_TAGS`. The report bypasses the need to have marked
the tags set to always trend tags as approved. (It does not mark them
as approved.)

* Adjust code to better satisfy upstream rubocop

This edits upstream code (as modified for this commit)

* Refactor to get always-trending tags from environment

And for simpler method stubbing in testing.

The original form of the commit forced a single particular tag to
trend. @kibigo correctly suggested to handle this generically.

* [DRAFT, TODO] Add tests for new always-trending tags code

* Refactor Api::V1::Trends::TagsController for testing

* Add tests for always-trending tags

* Adopt @kibigo's suggestions

As per
https://gitlab.com/kibicat/mastodon/-/merge_requests/7#note_931604764
and
https://gitlab.com/kibicat/mastodon/-/merge_requests/7#note_931607104