CI: Support larger diff artifacts
(Much too large) MRs can theoretically exceed the artifact size limit
for dotenv files; this ought to be more flexible for us.

A previous attempt just invoked deno as needed per stage, but this
involved installing deno as needed in non-deno images, and would be
ineffecient if we ever need the other diffs in any stage.