Add frontend support for inline spoiler text
To get around the fact that React manages the contents of statuses
through HTML injection, this commit adds a Redux action/reducer for
dynamically mutating the content of a status in the store. This
approach has some flaws:

- It’s messy as fuck, and

- If the status is reloaded for whatever reason (i·e it is received
  again through the A·P·I), the contents won’t match and there will be
  *another* (unnecessary) status body update (which will close any
  open spoilers in the status).

Still, I think this is the best that can be done with Mastodon’s
current architecture.