Support emojos in inline spoiler texts
Right now, this requires re·emojifying the spoiler texts every time one
is opened or closed. There is probably a significant more efficient way
of doing this, but this is easiest conceptually and from a maintenance
standpoint (and emojification is a fast enough process that I’m really
not worried about efficiency here).