Fix inline spoiler text in modals
This requires “connecting” `<StatusContent>`, which isn’t “great” but
is basically necessary at this point. On the plus side, this makes the
code a bit DRYer.

Due to the way modals work, updating status contents doesn’t actually
trigger a React rerender. This is actually fine, because we (have to)
update status contents live anyway, but it does require ensuring that
appropriate event listeners get attached so that the spoiler can be
toggled again.

To keep listeners from being added multiple times (which wouldn’t have
a negative effect but could be a performance leak), elements with
listeners are now tracked in a `WeakSet`.