Generate app icons before precomp
Squashed commit of
https://gitlab.com/kibicat/mastodon/-/merge_requests/30 and
https://gitlab.com/kibicat/mastodon/-/merge_requests/31

This causes :blobcat_sleep: to Actually be the favicon now; more
generally, it uses app/javascript/images/logo.png (ideally we would use
SVG---TODO: do that) as the basis fur the front end favicon and app
icons, which are now generated at build time.

This is arguably suboptimal as we purrform redundant work now on every
build, but I think that is an acceptable tradeoff. It's a small amount
of work and we do not exactly build the code very often.

I would prefer to not have to add an extra command to our build
scripts...

This removes app/javascript/icons and public/favicon.ico from repo.
Naturally.